refactor: move brc about roadmap page into sites/brc-analytics (#1637)
